requirements.txt
Django==1.11
mysqlclient==1.3.10
djangorestframework==3.6.3
django-rest-swagger==2.1.2
Faker==0.7.17
